datas

+++ stm32-cmake-blinking-led/source/main/inc/main/led_init.h
@@ -100,4 +100,4 @@
 #define  GPIO_PULLUP        0x00000001u   /*!< Pull-up activation                  */
 #define  GPIO_PULLDOWN      0x00000002u   /*!< Pull-down activation                */
 #endif
-int *init_led(void);
+uint16_t *init_led(void);
